pyproject.toml
[project]
name = "langchain-pra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.5.5",
    "click>=8.3.1",
    "langchain>=1.2.12",
    "langchain-chroma>=1.1.0",
    "langchain-core>=1.2.19",
    "langchain-google-genai>=4.2.1",
    "langchain-text-splitters>=1.1.1",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = [
    "deptry>=0.24.0",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "ruff>=0.15.6",
]

[tool.ruff]
target-version = "py314"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.14"
strict = true
ignore_missing_imports = true

[tool.deptry]
ignore = ["DEP002"]
